requirements.txt
camel-ai
pydantic
loguru
python-dotenv
numpy
requests
librosa
opencv-python
sqlalchemy
psycopg2-binary
aiosqlite
soundfile
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
pydantic[email]
chromadb
pyyaml
pydantic-settings
